@@ -2,11 +2,15 @@
   "name": "capital-server",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "npx tsc && npx tsc-alias",
-    "start": "npm run build && NODE_ENV=production ts-node api/index.ts",
-    "dev": "npm run build && nodemon -L api/index.ts",
-    "lint": "npx eslint '**/*.ts' --fix --ignore-pattern '**/build/'"
+    "build": "npx tsc && npx tsc-alias --resolve-full-paths",
+    "start": "npm run build && NODE_ENV=production node --no-warnings=ExperimentalWarning build/api/index.js",
+    "dev": "nodemon -L --watch '**/*.ts' --ext ts --exec 'npm run build && node build/api/index.js'",
+    "lint": "npx eslint '**/*.ts' --ignore-pattern '**/build/'",
+    "lint:fix": "npx eslint '**/*.ts' --fix --ignore-pattern '**/build/'",
+    "typecheck": "npx tsc --noEmit",
+    "test": "npx jest tests/* --collect-coverage --detectOpenHandles --verbose"
   },
   "keywords": [],
   "author": "Jeffrey Cordero",
@@ -33,24 +37,30 @@
     "module-alias": "^2.2.3",
     "morgan": "~1.9.1",
     "pg": "^8.13.3",
+    "pgsql-ast-parser": "^12.0.1",
     "redis": "^4.6.12",
+    "tsc-alias": "^1.8.16",
+    "typescript": "^5.9.3",
     "util": "^0.12.5",
     "winston": "^3.17.0",
     "zod": "^3.22.2"
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
-    "@types/compression": "^1.7.5",
+    "@types/compression": "^1.8.1",
     "@types/cookie-parser": "^1.4.8",
     "@types/cors": "^2.8.17",
     "@types/crypto-js": "^4.2.2",
     "@types/express": "^5.0.0",
     "@types/express-rate-limit": "^5.1.3",
+    "@types/jest": "^29.5.14",
     "@types/jsonwebtoken": "^9.0.7",
+    "@types/module-alias": "^2.0.4",
     "@types/morgan": "^1.9.9",
-    "@types/node": "^22.10.4",
+    "@types/node": "^22.18.12",
     "@types/pg": "^8.11.11",
     "@types/serve-index": "^1.9.4",
+    "@types/supertest": "^6.0.2",
     "@types/winston": "^2.4.4",
     "@types/xml2js": "^0.4.14",
     "@typescript-eslint/eslint-plugin": "^8.23.0",
@@ -62,12 +72,12 @@
     "eslint-plugin-import-newlines": "^1.4.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-spellcheck": "^0.0.20",
+    "jest": "^29.7.0",
     "nodemon": "^3.1.9",
-    "ts-node": "^10.9.2",
-    "tsc-alias": "^1.8.15",
-    "typescript": "^5.8.3"
+    "ts-jest": "^29.4.5",
+    "ts-node": "^10.9.2"
   },
   "_moduleAliases": {
-    "@": "./"
+    "@": "./build"
   }
 }
