@@ -12,25 +12,16 @@ import {
    Stack,
    Typography
 } from "@mui/material";
-import { userSchema } from "capital/user";
+import { loginSchema } from "capital/user";
 import clsx from "clsx";
 import { useState } from "react";
 import { Controller, useForm } from "react-hook-form";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router";
-import { z } from "zod";
 
 import Callout from "@/components/global/callout";
 import { sendApiRequest } from "@/lib/api";
 import { handleValidationErrors } from "@/lib/validation";
-
-/**
- * Login schema with username and password fields
- */
-const loginSchema = z.object({
-   username: userSchema.innerType().shape.username,
-   password: userSchema.innerType().shape.password
-});
 
 /**
  * Login page component with form validation and authentication
@@ -65,7 +56,9 @@ export default function Login(): React.ReactNode {
    };
 
    return (
-      <Container className = "center">
+      <Container
+         className = "center"
+      >
          <Callout
             sx = { { width: "100%" } }
          >
@@ -92,7 +85,10 @@ export default function Login(): React.ReactNode {
                      Login
                   </Typography>
                </Stack>
-               <form onSubmit = { handleSubmit(onSubmit) }>
+               <form
+                  aria-label = "Login Form"
+                  onSubmit = { handleSubmit(onSubmit) }
+               >
                   <Stack
                      direction = "column"
                      spacing = { 1.5 }
@@ -102,7 +98,9 @@ export default function Login(): React.ReactNode {
                         name = "username"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.username) }>
+                              <FormControl
+                                 error = { Boolean(errors.username) }
+                              >
                                  <InputLabel htmlFor = "username">
                                     Username
                                  </InputLabel>
@@ -111,6 +109,7 @@ export default function Login(): React.ReactNode {
                                     autoComplete = "username"
                                     autoFocus = { true }
                                     id = "username"
+                                    inputProps = { { "data-testid": "username" } }
                                     label = "Username"
                                     type = "text"
                                     value = { field.value || "" }
@@ -127,7 +126,9 @@ export default function Login(): React.ReactNode {
                         name = "password"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.password) }>
+                              <FormControl
+                                 error = { Boolean(errors.password) }
+                              >
                                  <InputLabel htmlFor = "password">
                                     Password
                                  </InputLabel>
@@ -143,6 +144,7 @@ export default function Login(): React.ReactNode {
                                        />
                                     }
                                     id = "password"
+                                    inputProps = { { "data-testid": "password" } }
                                     label = "Password"
                                     type = { showPassword ? "text" : "password" }
                                     value = { field.value || "" }
@@ -157,6 +159,7 @@ export default function Login(): React.ReactNode {
                      <Button
                         className = "btn-primary"
                         color = "primary"
+                        data-testid = "submit-button"
                         fullWidth = { true }
                         loading = { isSubmitting }
                         loadingPosition = "start"
@@ -174,9 +177,11 @@ export default function Login(): React.ReactNode {
                         Don&apos;t have an account?{ " " }
                         <Link
                            color = "primary"
+                           data-testid = "register-link"
                            fontWeight = "bold"
-                           id = "register"
+                           id = "register-link"
                            onClick = { () => navigate("/register") }
+                           role = "link"
                            underline = "none"
                         >
                            Register
