@@ -96,7 +96,11 @@ export default function Register(): React.ReactNode {
                      Register
                   </Typography>
                </Stack>
-               <form onSubmit = { handleSubmit(onSubmit) }>
+               <form
+                  aria-label = "Register Form"
+                  noValidate = { true }
+                  onSubmit = { handleSubmit(onSubmit) }
+               >
                   <Stack
                      direction = "column"
                      spacing = { 1.5 }
@@ -106,7 +110,9 @@ export default function Register(): React.ReactNode {
                         name = "name"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.name) }>
+                              <FormControl
+                                 error = { Boolean(errors.name) }
+                              >
                                  <InputLabel htmlFor = "name">
                                     Name
                                  </InputLabel>
@@ -115,6 +121,7 @@ export default function Register(): React.ReactNode {
                                     autoComplete = "name"
                                     autoFocus = { true }
                                     id = "name"
+                                    inputProps = { { "data-testid": "name" } }
                                     label = "Name"
                                     type = "text"
                                     value = { field.value || "" }
@@ -131,7 +138,9 @@ export default function Register(): React.ReactNode {
                         name = "birthday"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.birthday) }>
+                              <FormControl
+                                 error = { Boolean(errors.birthday) }
+                              >
                                  <InputLabel
                                     htmlFor = "birthday"
                                     shrink = { true }
@@ -146,6 +155,7 @@ export default function Register(): React.ReactNode {
                                     slotProps = {
                                        {
                                           htmlInput: {
+                                             "data-testid": "birthday",
                                              min: minDate,
                                              max: maxDate
                                           },
@@ -174,7 +184,9 @@ export default function Register(): React.ReactNode {
                         name = "username"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.username) }>
+                              <FormControl
+                                 error = { Boolean(errors.username) }
+                              >
                                  <InputLabel htmlFor = "username">
                                     Username
                                  </InputLabel>
@@ -182,6 +194,7 @@ export default function Register(): React.ReactNode {
                                     { ...field }
                                     autoComplete = "none"
                                     id = "username"
+                                    inputProps = { { "data-testid": "username" } }
                                     label = "Username"
                                     type = "text"
                                     value = { field.value || "" }
@@ -198,7 +211,9 @@ export default function Register(): React.ReactNode {
                         name = "password"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.password) }>
+                              <FormControl
+                                 error = { Boolean(errors.password) }
+                              >
                                  <InputLabel htmlFor = "password">
                                     Password
                                  </InputLabel>
@@ -214,6 +229,7 @@ export default function Register(): React.ReactNode {
                                        />
                                     }
                                     id = "password"
+                                    inputProps = { { "data-testid": "password" } }
                                     label = "Password"
                                     type = { showPassword ? "text" : "password" }
                                     value = { field.value || "" }
@@ -230,7 +246,9 @@ export default function Register(): React.ReactNode {
                         name = "verifyPassword"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.verifyPassword) }>
+                              <FormControl
+                                 error = { Boolean(errors.verifyPassword) }
+                              >
                                  <InputLabel htmlFor = "verifyPassword">
                                     Verify Password
                                  </InputLabel>
@@ -246,6 +264,7 @@ export default function Register(): React.ReactNode {
                                        />
                                     }
                                     id = "verifyPassword"
+                                    inputProps = { { "data-testid": "verifyPassword" } }
                                     label = "Verify Password"
                                     type = { showVerifyPassword ? "text" : "password" }
                                     value = { field.value || "" }
@@ -262,7 +281,9 @@ export default function Register(): React.ReactNode {
                         name = "email"
                         render = {
                            ({ field }) => (
-                              <FormControl error = { Boolean(errors.email) }>
+                              <FormControl
+                                 error = { Boolean(errors.email) }
+                              >
                                  <InputLabel htmlFor = "email">
                                     Email
                                  </InputLabel>
@@ -270,6 +291,7 @@ export default function Register(): React.ReactNode {
                                     { ...field }
                                     autoComplete = "email"
                                     id = "email"
+                                    inputProps = { { "data-testid": "email" } }
                                     label = "email"
                                     type = "email"
                                     value = { field.value || "" }
@@ -284,6 +306,7 @@ export default function Register(): React.ReactNode {
                      <Button
                         className = "btn-primary"
                         color = "primary"
+                        data-testid = "submit-button"
                         fullWidth = { true }
                         loading = { isSubmitting }
                         loadingPosition = "start"
@@ -301,8 +324,11 @@ export default function Register(): React.ReactNode {
                         Already have an account?{ " " }
                         <Link
                            color = "primary"
+                           data-testid = "login-link"
                            fontWeight = "bold"
+                           id = "login-link"
                            onClick = { () => navigate("/login") }
+                           role = "link"
                            underline = "none"
                         >
                            Login
