@@ -1,4 +1,4 @@
-import { ServerResponse } from "capital/server";
+import { HTTP_STATUS, ServerResponse } from "capital/server";
 import { Response } from "express";
 import { SafeParseReturnType } from "zod";
 
@@ -7,7 +7,7 @@ import { removeCacheValue } from "@/lib/redis";
 import { sendErrors, sendSuccess } from "@/lib/response";
 
 /**
- * Formats validation errors with a 400 status code based on Zod schema results.
+ * Formats validation errors with `HTTP_STATUS.BAD_REQUEST` based on Zod schema results.
  *
  * @param {SafeParseReturnType<any, any> | null} fields - Zod validation results or null
  * @param {Record<string, string>} [errors] - Optional prepared error messages (used when fields is null)
@@ -22,7 +22,7 @@ export function sendValidationErrors(
       const errors = fields.error?.flatten().fieldErrors || {};
 
       return {
-         code: 400,
+         statusCode: HTTP_STATUS.BAD_REQUEST,
          errors: Object.fromEntries(
             Object.entries(errors as Record<string, string[]>).map(([field, errors]) => [
                field, errors?.at(0) || "Unknown error"
@@ -32,7 +32,7 @@ export function sendValidationErrors(
    } else {
       // Use predefined validation errors
       return {
-         code: 400,
+         statusCode: HTTP_STATUS.BAD_REQUEST,
          errors: errors || {}
       };
    }
@@ -41,15 +41,15 @@ export function sendValidationErrors(
 /**
  * Creates a structured service response with specified status code and data.
  *
- * @param {number} code - HTTP status code
+ * @param {number} statusCode - HTTP status code
  * @param {any} [data] - Optional response data
  * @param {Record<string, string>} [errors] - Optional error details
  * @returns {ServerResponse} Formatted server response
  */
-export function sendServiceResponse(code: number, data?: any, errors?: Record<string, string>): ServerResponse {
+export function sendServiceResponse(statusCode: number, data?: any, errors?: Record<string, string>): ServerResponse {
    return {
-      code: code,
-      data: data ?? undefined,
+      statusCode,
+      data: data || undefined,
       errors: errors || undefined
    };
 }
@@ -60,35 +60,35 @@ export function sendServiceResponse(code: number, data?: any, errors?: Record<st
  * @param {Response} res - Express response object
  * @param {Function} serviceMethod - Async function containing the service logic
  */
-export const submitServiceRequest = async(
+export async function submitServiceRequest(
    res: Response,
    serviceMethod: () => Promise<ServerResponse>
-): Promise<void> => {
+): Promise<void>  {
    try {
       const result: ServerResponse = await serviceMethod();
 
-      if (result.code === 200 || result.code === 201 || result.code === 204) {
+      if (result.statusCode === HTTP_STATUS.OK || result.statusCode === HTTP_STATUS.CREATED || result.statusCode === HTTP_STATUS.NO_CONTENT || result.data?.refreshable) {
          // Success response
-         return sendSuccess(res, result.code, result.data ?? undefined);
+         return sendSuccess(res, result.statusCode, result.data ?? undefined);
       } else {
          // Error response
-         return sendErrors(res, result.code, result.errors);
+         return sendErrors(res, result.statusCode, result.errors);
       }
    } catch (error: any) {
       // Log unexpected errors
       logger.error(error.stack);
 
-      return sendErrors(res, 500, { server: "Internal Server Error" });
+      return sendErrors(res, HTTP_STATUS.INTERNAL_SERVER_ERROR, { server: "Internal Server Error" });
    }
-};
+}
 
 /**
  * Helper function to send a successful update response after clearing a cache key for strong consistency.
  *
  * @param {string} key - Cache key
- * @returns {Promise<ServerResponse>} A server response of `204` with no content
+ * @returns {Promise<ServerResponse>} A server response of `HTTP_STATUS.NO_CONTENT` with no content
  */
-export const clearCacheAndSendSuccess = (key: string): ServerResponse => {
+export function clearCacheAndSendSuccess(key: string): ServerResponse {
    removeCacheValue(key);
-   return sendServiceResponse(204);
-};
+   return sendServiceResponse(HTTP_STATUS.NO_CONTENT);
+}
